# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SPM API
  version: 1.0.0
  description: Senior Project Management system API
servers:
  - url: /api
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorMessage:
      type: object
      properties:
        message:
          type: string
...
    GroupMember:
      type: object
      properties:
        studentId:
          type: string
        role:
          type: string
          enum: [TEAM_LEADER, MEMBER]
        joinedAt:
          type: string
          format: date-time
...
    GroupDetailResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        groupName:
          type: string
        termId:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        jiraSpaceUrl:
          type: string
          nullable: true
        jiraProjectKey:
          type: string
          nullable: true
        jiraBound:
          type: boolean
        githubOrgName:
          type: string
          nullable: true
        githubBound:
          type: boolean
        advisorId:
          type: string
          format: uuid
          nullable: true
        members:
          type: array
          items:
            $ref: '#/components/schemas/GroupMember'
...
    RespondInvitationRequest:
      type: object
      required: [accept]
      properties:
        accept:
          type: boolean
    DeclineInvitationResponse:
      type: object
      properties:
        invitationId:
          type: string
          format: uuid
        status:
          type: string
...
security:
  - bearerAuth: []
paths:
...
  /invitations/{invitationId}:
    delete:
      tags: [Student]
      summary: Cancel a pending invitation
      parameters:
        - name: invitationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Invitation cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeclineInvitationResponse'
        '400':
          description: Invitation is no longer pending
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '403':
          description: Caller is not the Team Leader
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '404':
          description: Invitation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
...
  /invitations/{invitationId}/respond:
    patch:
      tags: [Student]
      summary: Respond to an invitation
      parameters:
        - name: invitationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RespondInvitationRequest'
      responses:
        '200':
          description: Invitation processed successfully
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/GroupDetailResponse'
                  - $ref: '#/components/schemas/DeclineInvitationResponse'
        '400':
          description: Invitation is no longer pending; or group status is TOOLS_BOUND or higher (roster lock blocks student accept)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '403':
          description: Invitation does not belong to the user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '404':
          description: Invitation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
...